requirements.txt
# Green AI Wireless Prototype - Requirements
# RF Modulation Classification with Green AI Compression Techniques

# Deep Learning Framework
torch==2.2.2

# Model Deployment & Optimization
onnx==1.16.0
onnxruntime==1.18.0

# Scientific Computing
numpy==1.26.4
scipy==1.12.0

# Visualization & Analysis
matplotlib==3.8.4

# Utilities
tqdm==4.66.4           # Progress bars
pyyaml==6.0.1          # YAML configuration files